reorganize files and remove unused api_caller class. made spec improvements and increased code coverage. one code modification to fix error messaging, where order.errors is never null so the existing logic was flawed.
