+++ lib/spree_ccavenue.rb
@@ -1,5 +1,3 @@
 require 'spree_core'
 require 'spree_ccavenue/engine'
-require 'ccavenue-sdk'
-require 'aes_crypter'
-require 'api_caller'
+require 'ccavenue_api/sdk'
