+++ spec/controllers/spree/ccavenue_controller_spec.rb
@@ -1,10 +1,8 @@
 require 'spec_helper'
 describe Spree::CcavenueController, :type => :controller do
+
   let(:order) { FactoryGirl.create(:order_with_totals) }
-
-  let!(:ccavenue_gw) { Spree::Gateway::Ccavenue.create!(:name        => 'ccavenue test gw',
-                                                     environment: Rails.env) }
-
+  let!(:ccavenue_gw) { Spree::Gateway::Ccavenue.create!(name: "ccavenue test gw", environment: Rails.env) }
   let(:merchant_id) { '1234' }
   let(:enc_key) { 'test#test' }
   let(:access_code) { '1234' }
@@ -28,13 +26,12 @@ describe Spree::CcavenueController, :type => :controller do
     Spree::Ccavenue::Transaction.create!(tracking_id: '123', auth_desc: 'Aborted', ccavenue_order_number: order.number, ccavenue_amount: order.total.to_s)
   }
   let(:ccavenue_response) { double('ccavenue_response') }
-
   let(:encResp) { '123' }
 
   before do
     order.state = 'payment'
     order.save!
-    allow(controller).to receive(:payment_method).and_return(ccavenue_gw)
+    allow(Spree::PaymentMethod).to receive(:find).and_return(ccavenue_gw)
     allow(ccavenue_gw).to receive(:provider).and_return(ccavenue_provider)
   end
 
@@ -106,7 +103,7 @@ describe Spree::CcavenueController, :type => :controller do
 
     context "when current_order is nil" do
       before do
-        allow(controller).to receive(:ccavenue_transaction).and_return(successful_ccavenue_transaction)
+        allow(Spree::Ccavenue::Transaction).to receive(:find).and_return(successful_ccavenue_transaction)
         expect(controller).to receive(:current_order).at_least(:once).and_return(nil)
       end
 
@@ -119,7 +116,7 @@ describe Spree::CcavenueController, :type => :controller do
 
     context "when the transaction does not exist on the store side" do
       before do
-        expect(controller).to receive(:ccavenue_transaction).and_return(nil)
+        allow(Spree::Ccavenue::Transaction).to receive(:find).and_return(nil)
       end
 
       it "redirects to cart with the correct flash message" do
@@ -129,18 +126,55 @@ describe Spree::CcavenueController, :type => :controller do
       end
     end
 
-
     context "on successful ccavenue transaction" do
       before do
         expect(controller).to receive(:current_order).at_least(:once).and_return(order)
-        allow(controller).to receive(:ccavenue_transaction).and_return(successful_ccavenue_transaction)
+        allow(Spree::Ccavenue::Transaction).to receive(:find).and_return(successful_ccavenue_transaction)
       end
+
       context "when the order is successfully completed" do
-        it 'redirects to order completion route' do
+
+        it "redirects to order completion route" do
           do_post
           expect(response).to redirect_to spree.order_path(order)
           expect(flash[:notice]).to eq(Spree.t('ccavenue.order_processed_successfully'))
         end
+
+      end
+
+      context "when the order is not successfully completed" do
+
+        before do
+          expect(order).to receive(:complete?).and_return(false)
+        end
+
+        it "redirects to new order checkout state path" do
+          do_post
+          expect(response).to redirect_to spree.checkout_state_path(order.state)
+        end
+
+        context "order has errors" do
+
+          before do
+            allow(order).to receive(:errors).and_return(double("errors", full_messages: ["the_error"]).as_null_object)
+          end
+
+          it "shows errors" do
+            do_post
+            expect(flash[:error]).to eq "the_error"
+          end
+
+        end
+
+        context "order has no errors" do
+
+          it "shows generic error" do
+            do_post
+            expect(flash[:error]).to eq Spree.t("ccavenue.generic_failed")
+          end
+
+        end
+
       end
 
       context "when the inventory goes low" do
@@ -148,6 +182,7 @@ describe Spree::CcavenueController, :type => :controller do
           expect(order).to receive(:next).and_raise
           expect(controller).to receive(:out_of_stock_error).and_return(true)
         end
+
         context "and the void call succeeds" do
           before do
             expect(controller).to receive(:void_payment).and_return(true)
@@ -158,19 +193,34 @@ describe Spree::CcavenueController, :type => :controller do
             expect(flash[:error]).to eq(Spree.t('ccavenue.checkout_low_inventory_after_payment_warning'))
           end
         end
+
         context "and the void call fails" do
-          it "redirects with appropriate flash message when void fails" do
+
+          it "redirects with appropriate flash message" do
             expect(controller).to receive(:void_payment).and_return(false)
             do_post
             expect(flash[:error]).to eq(Spree.t('ccavenue.refund_api_call_failed'))
             expect(response).to redirect_to spree.cart_path
           end
+
+          context "due to an exception" do
+
+            before do
+              allow_any_instance_of(Spree::Payment).to receive(:void_transaction!) { raise StandardError}
+            end
+
+            it "returns generic message" do
+              do_post
+              expect(flash[:error]).to eq Spree.t('ccavenue.refund_api_call_failed')
+            end
+          end
+
         end
       end
 
       context "when the order is changed" do
         before do
-          allow(controller).to receive(:ccavenue_transaction).and_return(changed_ccavenue_transaction)
+          allow(Spree::Ccavenue::Transaction).to receive(:find).and_return(changed_ccavenue_transaction)
           expect(controller).to receive(:void_payment).and_return(true)
         end
 
@@ -182,11 +232,10 @@ describe Spree::CcavenueController, :type => :controller do
       end
     end
 
-
-    context 'when payment is aborted at ccavenue' do
+    context "when payment is aborted at ccavenue" do
       before do
         expect(controller).to receive(:current_order).at_least(:once).and_return(order)
-        allow(controller).to receive(:ccavenue_transaction).and_return(aborted_ccavenue_transaction)
+        allow(Spree::Ccavenue::Transaction).to receive(:find).and_return(aborted_ccavenue_transaction)
         expect(order).to receive(:next).and_raise
         expect(controller).to receive(:out_of_stock_error).and_return(false)
         do_post
@@ -197,10 +246,10 @@ describe Spree::CcavenueController, :type => :controller do
       end
     end
 
-    context 'when payment fails at ccavenue' do
+    context "when payment fails at ccavenue" do
       before do
         expect(controller).to receive(:current_order).at_least(:once).and_return(order)
-        allow(controller).to receive(:ccavenue_transaction).and_return(failed_ccavenue_transaction)
+        allow(Spree::Ccavenue::Transaction).to receive(:find).and_return(failed_ccavenue_transaction)
         expect(order).to receive(:next).and_raise
         expect(controller).to receive(:out_of_stock_error).and_return(false)
         do_post
